document_collection/requirements.txt
# Core dependencies
pydantic>=2.0.0
requests>=2.31.0
click>=8.1.0
pyyaml>=6.0
pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.0
# MCP will be added when available

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
mypy>=1.5.0
ruff>=0.0.287
pre-commit>=3.4.0
